+++ packages/oauth-client/src/CognitoClient.ts
@@ -46,10 +46,11 @@ export class CognitoClient extends OAuthClient {
   };
 
   /**
-   * Query the Cognito API for the user object associated with an access token.
+   * Query the API for the user object associated with an access token.
    *
    * @param {Object} params
    * @param {string} params.token - The access token for Authorization header
+   * @param {string} [params.xRequestId] - a string to help identify the request
    * @returns {Promise<Object>} The user object (see example)
    *
    * @example
@@ -67,7 +68,7 @@ export class CognitoClient extends OAuthClient {
     token: string,
     xRequestId?: string,
   }) {
-    const { token, xRequestId } = params;
+    const { token, xRequestId } = params || {};
     if (!token) throw new TypeError('token is required');
     const headers = xRequestId ? { 'X-Request-Id': xRequestId } : undefined;